PortalService/requirements.txt
pymongo==3.11.0
flask
flask-restx
mailjet-rest==1.3.4
requests
python-dotenv
dnspython==2.0.0
opentelemetry-exporter-otlp
opentelemetry-distro
prometheus-client
uwsgi